indexer/pyproject.toml
[project]
name = "ragstral-indexer"
version = "0.1.0"
description = "RAG indexing pipeline for GitHub repositories"
requires-python = ">=3.9"
dependencies = [
    "mistralai>=1.2.0",
    "mistral-common>=1.5.0",
    "pinecone>=5.0.0",
    "langchain>=0.2.0",
    "langchain-text-splitters>=0.2.0",
    "requests>=2.31.0",
    "numpy>=1.24.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-settings>=2.0.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "huggingface-hub>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["services"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.uv]
dev-dependencies = [
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
] 